Fixed a security bug with path handling and file deletion
Previously, the full path of the file we wanted to delete was passed
in via an AJAX delete call.  However, there was no verification of the
filename, which could allow a malicious user to delete any file that
was writeable by the web server.  This patch stores all filenames in
session and attempts to verify that this user is allowed to delete them
before deleting them.

This fix doesn't totally prevent a malicious user from attempting to
delete files that don't belong to the user with the same file name.
If a malicious user uploads a file with the same names as someone else
on the same day, even if the system prevents him from uploading the
file he may still be allowed to delete that other file.