Setup hashing of filenames to work on preventing a DOS attack by a malicious user attempting to delete another user's file
Filenames are hashed using sha1 (to be fast, but a bit more secure
than md5).  The hash consists of the filename (not path), user id
and the current time.

The default for this change is setup to be backwards compatible
in case other systems rely upon a certain length filename or
some other method of parsing the filename.